requirements.txt
Flask==3.0.0
Flask-PyMongo==2.3.0
Werkzeug==3.0.0
python-dotenv==1.0.1
gunicorn
dnspython
pymongo
